Adapt Express Server for Vercel
