+++ server/package.json
@@ -2,10 +2,11 @@
   "name": "server",
   "version": "1.0.0",
   "description": "My server",
-  "main": "index.js",
+  "main": "api/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon index.js"
+    "dev": "nodemon api/index.js",
+    "vercel-build": "echo hello"
   },
   "author": "",
   "license": "ISC",
@@ -18,8 +19,10 @@
     "mongoose": "^8.8.1",
     "multer": "^1.4.5-lts.1",
     "multer-storage-cloudinary": "^4.0.0",
-    "nodemon": "^3.1.7",
     "sanitize-html": "^2.14.0",
     "socket.io": "^4.8.1"
+  },
+  "devDependencies": {
+    "nodemon": "^3.1.7"
   }
 }
