+++ vercel.json
@@ -1,0 +1,25 @@
+{
+  "version": 2,
+  "buildCommand": "cd src && npm run build",
+  "outputDirectory": "src/dist",
+  "builds": [
+    {
+      "src": "server/api/index.js",
+      "use": "@vercel/node"
+    },
+    {
+      "src": "src/**",
+      "use": "@vercel/static"
+    }
+  ],
+  "routes": [
+    {
+      "src": "/api/(.*)",
+      "dest": "server/api/index.js"
+    },
+    {
+      "src": "/(.*)",
+      "dest": "src/$1"
+    }
+  ]
+}
